requirements.txt
eel>=0.16.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
feedparser>=6.0.10
aiohttp>=3.9.0
